-parait printf avec ajout du dprintf

+++ src/ft_option_s.c
@@ -6,7 +6,7 @@
 /*   By: rgermain <marvin@le-101.fr>                +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2018/11/12 12:38:05 by rgermain     #+#   ##    ##    #+#       */
-/*   Updated: 2018/12/01 15:43:00 by rgermain    ###    #+. /#+    ###.fr     */
+/*   Updated: 2018/12/01 16:21:41 by rgermain    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -81,7 +81,7 @@ int			*ft_params_smaj(t_valst *lst_va, char *str, int *tab_i, int index)
 	char		*new;
 
 	lst = ft_put_option(lst_va, str, tab_i[0], index);
-	new = (long)va_arg(lst_va->lst_copy, char*);
+	new = va_arg(lst_va->lst_copy, char*);
 	lst->af_nb = ft_abs(lst->af_nb);
 	if (new == NULL)
 		new = "(null)";
